% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPPDSurv-package.R
\docType{package}
\name{BayesPPDSurv-package}
\alias{BayesPPDSurv-package}
\title{Bayesian sample size determination using the power and normalized power prior for survival data}
\description{
The \pkg{BayesPPDSurv} (Bayesian Power Prior Design for Survival Data) package provides two categories of functions:
functions for Bayesian power/type I error calculation and functions for model fitting.
}
\references{
Chen, Ming-Hui, et al. "Bayesian design of noninferiority trials for medical devices using historical data." Biometrics 67.3 (2011): 1163-1170.
}
