% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPPDSurv-package.R
\docType{package}
\name{BayesPPDSurv-package}
\alias{BayesPPDSurv-package}
\title{Bayesian sample size determination using the power and normalized power prior for survival data}
\description{
The \pkg{BayesPPDSurv} (Bayesian Power Prior Design for Survival Data) package provides two categories of functions:
functions for Bayesian power/type I error calculation and functions for model fitting.
}
\references{
Ibrahim, J. G., Chen, M.-H. and Sinha, D. (2001). Bayesian Survival Analysis. New York: Springer Science & Business Media.

Psioda, M. A. and Ibrahim, J. G. (2019). Bayesian clinical trial design using historical data that inform the treatment effect. Biostatistics 20, 400–415.

Shen, Y., Psioda, M. A., and Joseph, J. G. (2023). BayesPPD: an R package for Bayesian sample size determination using the power and normalized power prior for generalized linear models. The R Journal, 14(4).
}
